requirements.txt
langgraph
spacy
transformers
torch
readability-lxml
faiss-cpu
chromadb
pandas
numpy
jinja2
streamlit
requests
tiktoken
sentence-transformers
# mcp[cli]
python-dotenv
fastmcp